Fix constant_pad_nd_mps bug when pad is empty (#161149)
Fixes #161066

There is a size check here, which causes the error.
https://github.com/pytorch/pytorch/blob/c8a84de242accf2b0d5934ca0fb4980985828f35/aten/src/ATen/native/mps/operations/Pad.mm#L39-L40

If the argument `pad` is empty, it will return the cloned tensor on CPU.

https://github.com/pytorch/pytorch/blob/c8a84de242accf2b0d5934ca0fb4980985828f35/aten/src/ATen/native/PadNd.cpp#L43-L64

Therefore, this PR fixes the empty padding argument error by checking the size first and returning a cloned tensor immediately if the padding size is 0.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/161149
Approved by: https://github.com/malfet