[Dynamo] Fix weakref.proxy error when `torch.compile` (#161508)
Fixes #159258

The error occurs when we attempt to create a weak reference from a weak reference proxy.
https://github.com/pytorch/pytorch/blob/f85906a67cc999250d7f3ad3b62b62a671aa099e/torch/_dynamo/guards.py#L2910-L2915

In fact, we shouldn't create a weak reference from another reference or proxy, as it would check in CPython.
https://github.com/python/cpython/blob/f60f8225ed146a8f9b5fbf1eeed3474782127ea8/Objects/weakrefobject.c#L410-L418

However, `__weakrefoffset__` is not equal to **0** when the `guarded_object` is in `weakref.ProxyTypes`, and it will wrongly create a weak reference for the `weakref.ProxyTypes`. I think this could be a bug from CPython, but we can prevent it by adding more weakref type checks (`weakref.ProxyTypes` contains `weakref.ProxyType` and `weakref.CallableProxyType`) here.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/161508
Approved by: https://github.com/Lucaskabela, https://github.com/anijain2305, https://github.com/malfet